div test 추가

+++ src/operator/Div.java
@@ -2,6 +2,6 @@ package operator;
 
 public class Div implements Calculator {
 	public double calculate(int num1, int num2) {
-		return num1/num2;
+		return (double)num1/num2;
 	}
 }
